Changed name of SDC & TCL scripts to match repository name
When user first clones the GitHub repository, it will create it in a directory with the same name as the repository, i.e. SIUE-DigitalDesign-CPU-IC. Well, Dr. Engel's scripts expect the TCL and SDC file to have this same name.

When we were debugging the simulation with Cadence, we thought the problem was the dashes in the name but it turned out to be something else. So I want to see if changing the name back will cause problems.