Using 2048 words (8kB) took FOREVER to synthesize
Never even got a repsonse it took so long. Switched to using 64 words (256 bytes) instead and it synthesizes in about 5-10 minutes. From here, we will do our testing and stuff to get it working and maybe let the simulation run overnight to get it working for 2048 words.

Another note is that we only need 2kB and not 8kB so I could reduce this to see if it takes less time.

Experimentation with fanout, timing and etc to get it working right. Now at 70ns period (14MHz) with about 14ns of slack in it. However, the slowest path is the adder, so this can be reduced via a carry lookahead adder.