Clean up and alter SDC file
Add set_load command to set the capacitance on the output pins. This is set to be 1000fF or 1pF, which is what Bryan recommended we do. This is not yet tested using Cadence but initial theory is that it should not matter because this is not the longest path.