Fix RAM to utilize positive edge of the clock to allow full period delay
The RAM was the bottleneck and I realized that it was because on write it was only allowed half of the period for the combination logic to complete because the read/write of the RAM occurred on the negative edge of the clock.

The negative edge of the clock was used for the RAM module because of the read operation. The read operation has to occur before the positive edge because that's when it will be handled by the CPU.

Got around this by using combination logic to assign the data value continuously.