Implement Dan's left/right shifter modules
Now that I fixed the worst-case timing path to the RAM module, the next shortest path is the shift registers actually.

There was an instance where it was the adder but now that we have a custom adder, I **think** that is fixed.

Maybe having these custom shifters will make it faster for us?