Dr. Engel enlightened me that the set_max_delay will not do what I want it to.
He said from his experience, the set_max_delay command has very little to do with the synthesis tool. Rather, he suggested that we play around with some of the parameters given to the RTL compiler and try to tell it to 'work harder on time than power and area'.