Experimenting with getting place & route to work
Current idea is to reduce the clock speed even more to 10MHz. Place a larger clock uncertainty so that the slack is even larger. I also removed the max_fanout and max_capacitance values completely. I saw some errors in Encounter about this, so I want to see if this is affecting things.