Commented out all edited code to the TCL scripts
Was able to get down to 0 geometry violation which is a large improvement. There was 36 connectivity violations but they are different from the dangling clock net wires.

Seems like I have made progress. I was unable to run LVS or DRC on this due to not being able to convert it to an IC for some reason. edi2ic was being weird. edi2sch worked fine however.